Delete useless file.
