+++ modules/basic/ds/grape_perfect_hash/string_view_vector.h
@@ -78,45 +78,6 @@ class StringViewVector {
     offsets_.swap(rhs.offsets_);
   }
 
-  template <typename IOADAPTOR_T>
-  void serialize(std::unique_ptr<IOADAPTOR_T>& writer) const {
-    size_t content_buffer_size = content_buffer().size();
-    CHECK(writer->Write(&content_buffer_size, sizeof(size_t)));
-    if (content_buffer_size > 0) {
-      CHECK(writer->Write(const_cast<char*>(content_buffer().data()),
-                          content_buffer_size * sizeof(char)));
-    }
-    size_t offset_buffer_size = offset_buffer().size();
-    CHECK(writer->Write(&offset_buffer_size, sizeof(size_t)));
-    if (offset_buffer_size > 0) {
-      CHECK(writer->Write(const_cast<size_t*>(offset_buffer().data()),
-                          offset_buffer_size * sizeof(size_t)));
-    }
-  }
-
-  template <typename IOADAPTOR_T>
-  void deserialize(std::unique_ptr<IOADAPTOR_T>& reader) {
-    size_t content_buffer_size;
-    CHECK(reader->Read(&content_buffer_size, sizeof(size_t)));
-    if (content_buffer_size > 0) {
-      content_buffer().resize(content_buffer_size);
-      CHECK(reader->Read(content_buffer().data(),
-                         content_buffer_size * sizeof(char)));
-    }
-    size_t offset_buffer_size;
-    CHECK(reader->Read(&offset_buffer_size, sizeof(size_t)));
-    if (offset_buffer_size > 0) {
-      offset_buffer().resize(offset_buffer_size);
-      CHECK(reader->Read(offset_buffer().data(),
-                         offset_buffer_size * sizeof(size_t)));
-    }
-  }
-
-  void serialize_to_mem(std::vector<char>& buf) const {
-    encode_vec(buffer_, buf);
-    encode_vec(offsets_, buf);
-  }
-
  private:
   std::vector<char> buffer_;
   std::vector<size_t> offsets_;
